Fixed false positives inside Reach
This false positive is caused by attacking a stationary entity and while hitting, you are jumping back as much as you can to increase what reach Kauri thinks you're using. This flaw is caused by the pastlocation system not aiding when an entity is stationary. So, I decided to check if the entity is stationary and just subtract the reach distance by the deltaXZ if the target is stationary.